+++ packages/storage/src/common.ts
@@ -5,14 +5,15 @@
 //    *   Licensed under the Apache License, Version 2.0 (the "License");           *
 //    *******************************************************************************
 
+export * from "./tabular/CachedTabularRepository";
+export * from "./tabular/InMemoryTabularRepository";
 export * from "./tabular/ITabularRepository";
 export * from "./tabular/TabularRepository";
-export * from "./tabular/InMemoryTabularRepository";
 
 export * from "./kv/IKvRepository";
+export * from "./kv/InMemoryKvRepository";
 export * from "./kv/KvRepository";
 export * from "./kv/KvViaTabularRepository";
-export * from "./kv/InMemoryKvRepository";
 
-export * from "./queue/IQueueStorage";
 export * from "./queue/InMemoryQueueStorage";
+export * from "./queue/IQueueStorage";