+++ packages/ai-provider/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @podley/ai-provider
 
+## 0.0.30
+
+### Patch Changes
+
+- Add CachedTabularRepository and browser only SharedInMemoryTabularRepository
+- Updated dependencies
+  - @podley/task-graph@0.0.30
+  - @podley/job-queue@0.0.30
+  - @podley/storage@0.0.30
+  - @podley/util@0.0.30
+  - @podley/ai@0.0.30
+
 ## 0.0.29
 
 ### Patch Changes