+++ packages/storage/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @podley/storage
 
+## 0.0.30
+
+### Patch Changes
+
+- Add CachedTabularRepository and browser only SharedInMemoryTabularRepository
+- Updated dependencies
+  - @podley/sqlite@0.0.30
+  - @podley/util@0.0.30
+
 ## 0.0.29
 
 ### Patch Changes